Add files via upload
This project takes any XML file and checks whether or not is it RSS version 2.0, and prompts the user for a name of an html file that it will create. If the XML document is of the correct type and version then it creates header tags for an HTML document, loops through each node storing data into an HTML table with links, and creates closing tags, so that the document can easily be transferred into a readable table format.